requirements.txt
CherryPy==3.2.2
pymongo==2.6.3
python-firebase==1.1
pyzmq==2.2.0
scikit-learn==0.11
